rename token.py to admin_api.py
